requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
compressai==1.2.8
contourpy==1.3.0
cycler==0.12.1
einops==0.8.1
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
grpcio==1.76.0
idna==3.11
importlib-metadata==8.7.1
importlib-resources==6.5.2
jinja2==3.1.6
kiwisolver==1.4.7
lpips==0.1.4
markdown==3.9
markupsafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.7.0
networkx==3.2.1
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
pybind11==3.0.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-msssim==1.0.0
pytz==2025.2
requests==2.32.5
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tomli==2.3.0
torch==2.5.1+cu121
torch-geometric==2.6.1
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
werkzeug==3.1.4
yarl==1.22.0
zipp==3.23.0